+++ console/src/components/Main/State.ts
@@ -21,6 +21,7 @@ export interface MainState {
       is_function_permissions_inferred: boolean;
       is_admin_secret_set: boolean;
       is_auth_hook_set: boolean;
+      is_remote_schema_permissions_enabled: boolean;
       is_jwt_set: boolean;
       experimental_features: string[];
       jwt: {
@@ -59,6 +60,7 @@ const defaultState: MainState = {
       is_function_permissions_inferred: true,
       is_admin_secret_set: false,
       is_auth_hook_set: false,
+      is_remote_schema_permissions_enabled: false,
       experimental_features: [],
       is_jwt_set: false,
       jwt: {