+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 - server: fix issue with parsing of remote schema list of input objects (fix #6584)
 - cli: cli-ext is now a native part of cli binary (no longer needed as a plugin)
 - console: add browse rows for mssql tables (#805)
-
+- console: remote schema permissions bug fixes (#439)
 
 ## v2.0.0-alpha.4